Organize Jobs and Directory in different files
